create foundation for the new webapp including authentication process
 - create web client authentication process
   - web client asks botify to generate a token (UUID)
   - botify creates a token that is not currently used and stores it
     together with the IP address of the web client
   - user copies the generated token
   - user enters connect command, preparing a session for the current
     guild, member and text channel
   - user DMs the bot the token, completing the prepared session with
     the token and its associated ip address
   - if the web client already has a token stored it will check if there
     is an active session for it before the above steps
   - if the web client already has a client stored it will check if the
     session is still valid before the above steps
 - create simple EventWaiter implementation
   - used by the ConnectCommand to await private message
 - refactor CommandContext
   - add ExecutionContext super class for common context items between
     commands and rest request
   - add ThreadContext to manage all contexts installed for the current
     thread
     - install RequestContext for rest requests giving access to the
       current HttpServletRequest
 - create EventHandlerPool: thread pool used by all event listeners
   - instead of using either a cached or fixed thread pool use a custom
     configuration that has a small fixed core pool size and and a
     larger maximum pool size of threads with a 5 minute idle time
 - refactor LoginCommand
   - improve output and delete login link message upon completion since
     it cannot be reused
 - add CompletableFutures utility class to manage exceptions thrown by
   CompletableFuture#whenComplete
 - Playable#getSource: return enum instead of String
 - extend MessageService: refactor sending private messages
 - create ClassDescriptorNode and use the same logic as for
   GroovyWhitelistInterceptor for InjectorService to select the correct
   Extractor if there is one for a subclass of another Extractor

+++ src/main/java/net/robinfriedli/botify/concurrent/LoggingThreadFactory.java
@@ -1,7 +1,7 @@
 package net.robinfriedli.botify.concurrent;
 
 import java.util.concurrent.ThreadFactory;
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
 import net.robinfriedli.botify.exceptions.handlers.LoggingExceptionHandler;
 import org.jetbrains.annotations.NotNull;
@@ -9,7 +9,7 @@ import org.jetbrains.annotations.NotNull;
 public class LoggingThreadFactory implements ThreadFactory {
 
     private final String poolName;
-    private final AtomicInteger threadNumber = new AtomicInteger(1);
+    private final AtomicLong threadNumber = new AtomicLong(1);
 
     public LoggingThreadFactory(String poolName) {
         this.poolName = poolName;
